requirement.txt
Flask
torch
Pillow